pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "winagent"
version = "0.1.0"
description = "Multimodal Windows desktop AI agent (Gemini 2.5 Pro + tkinter + pyautogui)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.7.0",
    "pyautogui>=0.9.54",
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "pydantic>=2.5.0",
    "pywin32>=306; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
winagent = "winagent.__main__:main"

[tool.setuptools.packages.find]
include = ["winagent*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
